RAG/requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
langchain-text-splitters
chromadb
python-dotenv
pypdf
unstructured
fastapi
uvicorn
